requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cohere==5.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
docstring-parser==0.15
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
fastavro==1.9.4
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
iniconfig==2.0.0
instructor==1.0.0
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lxml==5.2.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.16.1
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pypdf==4.1.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.9.4
types-requests==2.31.0.20240402
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
XlsxWriter==3.2.0
yarl==1.9.4
